% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{poly_add_columns}
\alias{poly_add_columns}
\title{Add orthogonal polynomial columns to a dataframe}
\usage{
poly_add_columns(.data, .col, degree = 1, prefix = NULL,
  scale_width = NULL)
}
\arguments{
\item{.data}{a dataframe}

\item{.col}{a bare column name}

\item{degree}{number of polynomial terms to add to the dataframe}

\item{prefix}{prefix for the names to add to the dataframe. default is the
name of \code{.col}.}

\item{scale_width}{optionally rescale the dataframe using \code{\link[=poly_rescale]{poly_rescale()}}.
default is not to performa any rescaling.}
}
\value{
the dataframe with additional columns of orthogonal polynomial terms
of \code{.col}
}
\description{
Add orthogonal polynomial columns to a dataframe
}
