% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polypoly-package.R
\docType{package}
\name{polypoly}
\alias{polypoly}
\title{polypoly: Helper functions for orthogonal polynomials}
\description{
This package provides helpful functions for orthogonal polynomials created by
\code{\link[stats:poly]{stats::poly()}}. These include plotting \code{\link[=poly_plot]{poly_plot()}}, tidying \code{\link[=poly_melt]{poly_melt()}},
rescaling \code{\link[=poly_rescale]{poly_rescale()}}, and manipulating a dataframe
\code{\link[=poly_add_columns]{poly_add_columns()}}.
}
\author{
Tristan Mahr
}
