% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{poly_melt}
\alias{poly_melt}
\title{Melt a polynomial matrix}
\usage{
poly_melt(x)
}
\arguments{
\item{x}{a matrix created by \code{\link[stats:poly]{stats::poly()}}}
}
\value{
a \code{\link[tibble:tibble]{tibble::tibble()}} with three columns: \code{observation}, \code{degree}, and
\code{value}.
}
\description{
Melt a polynomial matrix
}
