% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{poly_plot}
\alias{poly_plot}
\title{Plot a polynomial matrix}
\usage{
poly_plot(x)
}
\arguments{
\item{x}{a matrix created by \code{\link[stats:poly]{stats::poly()}}}
}
\value{
a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} plot of the degree terms from the matrix
}
\description{
Plot a polynomial matrix
}
