% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{poly_rescale}
\alias{poly_rescale}
\title{Rescale the range of a polynomial matrix}
\usage{
poly_rescale(x, scale_width = 1)
}
\arguments{
\item{x}{a matrix created by \code{\link[stats:poly]{stats::poly()}}}

\item{scale_width}{the desired range (max - min) for the first column of the matrix}
}
\value{
the rescale polynomial matrix (as a plain matrix with attributes removed)
}
\description{
Rescale the range of a polynomial matrix
}
\details{
This function strips away the \code{poly} class and some of the
attributes of the matrix. This is because those attributes no longer
describe the transformed matrix.
}
