% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polypoly-package.R
\docType{package}
\name{polypoly}
\alias{polypoly}
\alias{polypoly-package}
\title{polypoly: Helper functions for orthogonal polynomials}
\description{
polypoly: Helper functions for orthogonal polynomials
}
